requirements.txt
flask
kafka-python
psycopg2-binary
boto3
pillow